+++ nbrshell/exec_shell_script.py
@@ -29,6 +29,7 @@
 #
 
 from IPython.core.magic import (register_cell_magic, needs_local_scope)
+from IPython.display import Javascript, display 
 
 # import common functions
 from . import nbrshell_common as cmn
@@ -102,5 +103,8 @@ def exec_shell_script(line, script):
         print(cmd)
         print("========= script-end =============")
 
+    # add html element with id="id_pbrun_as_oracle" for CSS to pick up
+    display( Javascript('element.setAttribute("id", "id_exec_shell_script")') )
+    
     # remote execute
     cmn._remote_execute_stream_output(host, user, psw, cmd)